fix: prevent admin selt delete or demote

+++ app/user/src/admin.ts
@@ -48,9 +48,17 @@ export async function listUsers(req: Request, res: Response, next: NextFunction)
         return res.status(403).json(new FoilCTF_Error("Forbidden", 403));
 
     const dbUsers = await db
-        .select()
+        .select({
+            id: users.id,
+            username: users.username,
+            email: users.email,
+            role: users.role,
+            // teamName: users.teamName, // after merging teams
+            createdAt: users.createdAt,
+
+        })
         .from(users)
-        .where(search ? ilike(users.username, `${search}`) : undefined)
+        .where(search ? ilike(users.username, `${search}%`) : undefined)
         .limit(limit)
         .offset(limit * (page - 1));
 
@@ -69,16 +77,26 @@ export async function updateUserRole(req: Request, res: Response, next: NextFunc
     if (decodedUser.role !== "admin")
         return res.status(403).json(new FoilCTF_Error("Forbidden", 403));
 
+    if (decodedUser.username === target)
+        return res.status(403).json(new FoilCTF_Error("You can not demote yourself!", 403));
+
+    const allowedRolesList = ["admin", "user"];
+    if (!allowedRolesList.includes(target))
+        return res.status(403).json(new FoilCTF_Error("Role not allowed", 403));
+
     await db.update(users).set({ role: newRole }).where(eq(users.username, target));
     return res.status(200).json(new FoilCTF_Success("OK", 200));
 }
 
-export async function deleteUser(req: Request, res: Response) {
+export async function deleteUser(req: Request, res: Response) { // what if user is in a team or has friends
     const target = req.params.username as string;
 
 	const decodedUser = res.locals.user;
     if (decodedUser.role !== "admin")
         return res.status(403).json(new FoilCTF_Error("Forbidden", 403));
+
+    if (decodedUser.username === target)
+        return res.status(403).json(new FoilCTF_Error("You can not delete yourself!", 403));
 
     await db.delete(users).where(eq(users.username, target));
     return res.status(204).json(new FoilCTF_Success("No Content", 204));
